Add id0076, id0077, id0078 (fast!)

+++ src/id0031.c
@@ -9,11 +9,11 @@ int main(void)
 {
     int coins[] = { 1, 2, 5, 10, 20, 50, 100, 200 };
     clock_t start = clock();
-    int* partitions = restricted_partition_range(201, coins, 8);
+    long long* partitions = restricted_partition_range(201, coins, 8);
 
     euler_assert(partitions);
 
-    int result = partitions[200];
+    long long result = partitions[200];
 
     free(partitions);
